Don't abort if we're using a git version
When I was hacking away on my own temporary fork, I ran into
difficulty trying to use my git version of lineman with a project
because the command-line tools would abort due to a version mismatch.
This is because it was looking at the git:// url in my project's
package.json and comparing it to the semver.